syncx: sync.Cond的超时等待版，Cond.WaitWithContext(ctx)
